Handle multiple addresses properly when resending verification mails
When sending a verification mail the intended address can be supplied as a second parameter. In the case multiple unverified addresses are available, this will ensure the mail is send to the address the user intended, rather than the first one.

Supplying the email address as a parameter also allows verification mails to be sent to addresses which are already verified. As such, the error handling which turns every possible error that could occur when sending a mail into "Already verified" has been removed, and a separate check to prevent sending verification mails to verified addresses is added.